feat(webhook): email admins in case of failed webhooks (#3654)
Lago sends many webhooks to the user's system but sometimes the webhook
fails.
We retry the webhook 3 times, then it dies.

This PR introduce a new email sent to all admins when a webhook fails to
be delivered. Usually, it means there is something wrong on the
receiving end.

This email is sent as soon as one webhook dies but is not sent again for
an hour. If the receiving endpoint is down for a few minutes, you don't
get flooded!

~**The email is text only, not HTML.** This is an admin email that you
should very rarely receive. It's easier than HTML.~ It's HTML.

![CleanShot 2025-05-21 at 17 15
07@2x](https://github.com/user-attachments/assets/c321a4fa-09c5-4abb-a7e3-f412adfdf854)