feat(pricing_group_keys): Add new `pricing_group_keys` API attribute (#3681)
## Context

This PR follows https://github.com/getlago/lago-api/pull/3666 and is
part of an epic to add grouped_by feature on all charge models and not
only on the standard one like today.

## Description

This PR is the second step:
- It updates the GraphQL and the Rest APIs to accept and return the new
`pricing_group_keys` attribute
- It makes sure that the deprecated `grouped_by` attribute is still
accepted and returned
- It adapt the aggregation logic to use the `pricing_group_keys` or the
`grouped_by` attributes

For now the grouped aggregation remains remains whitelisted only on the
standard and dynamic charge model. This will remain like this until the
logic is applied to all charge model