fix: Use current attributes instead of activejob_traceable (#3794)
Unfortunately, we've seen that our current use of `activejob_traceable`
is not thread safe with Sidekiq 😢

With the addition of activity logs, we've seen that `CurrentContext` can
use wrong information (`membership_id` and `source`) because we're not
resetting values at the start of each job. Two jobs can sequentially run
on the same thread.

The goal of this PR is to not use `activejob_traceable` anymore (and
remove a dependency 🎉 ), but introduce
`ActiveSupport::CurrentAttributes` in Sidekiq jobs.

This middleware allows to:
- persist values from `CurrentContext` into the Sidekiq job
- restore those values before the job starts running
- reset values after the job finishes

```ruby
module Sidekiq
  module CurrentAttributes
    class Middleware
      def call(worker, job, queue)
        # Deserialize CurrentContext attributes into thread
        ::CurrentContext.set(job["current_attributes"]) do
          yield
        end
      ensure
        ::CurrentContext.reset_all
      end
    end
  end
end
```