feat(stripe): Prepare test suite to use Stripe 2025-04-30.basil (#3787)
This PR adds fixtures for `2025-04-30.basil`. I removed the forced
version for some call if the client version is set.

The fixtures were reorganize a little to make it easier for next
upgrades:

1. All webhook payload are under `webhooks/`.
2. The webhook event file name is the name of the webhook event (no
suffix or extra description)
3. The content is always the full event (with `evt_*` id included)
4. HTTP call response fixutres are suffixed with `_response` but can
have extra description in the name.

Payloads are copy pasted AS-IS from stripe and never edited.
If you need to update the payload for the purpose of your tests, for
example setting the `metadata.lago_payable_id` (very typical), use the
helper to modify the payload. **Don't edit the files.**

Next time we upgrade, we only have to copy the payload to a file.

```rb
let(:event_json) do
  get_stripe_fixtures("webhooks/setup_intent_succeeded.json") do |h|
    h[:data][:object][:customer] = "cus_666" if h[:data][:object][:customer].nil?
    h[:data][:object][:metadata] = metadata
  end
end
let(:stripe_customer) do
  create(:stripe_customer, payment_provider: stripe_provider, customer:, provider_customer_id: "cus_123")
end
```