fix(invoice-preview): Support preview for subscription during upgrade or downgrade (#3744)
## Context

Current logic for invoice preview does not take into account if existing
subscription is in the process of downgrade or upgrade. To generate
proper invoice preview we need to slightly adjust data so that
subscription state matches one of a new bill run.

## Description

Add validation for customer persistence. It improves UX by providing
more precise error message for some invalid params.
Add support for proper preview of subscription in a process of being
downgraded.