fix(netsuite): Do not create duplicate contacts in netsuite (#3801)
## Context

An edge case might occur when there's already a contact created in
Netsuite but there's also another job that is trying to create the same
contact. In that case another contact in netsuite was created (with new
external_id) but then the job failed when trying to save the duplicate
in the DB

## Description

This PR fixes creating duplicate contacts in
`IntegrationCustomers::NetsuiteService` by first checking if the record
exists in `integration_customers`table in which case it doesn't even
call the nango endpoint and returns the existing integration customer.