chore(filter): Add the flat_filters database view (#3774)
## Context

Today, we are not able to use the cache system for organization using
Clickhouse event store, because we are not able to easily identify if an
event matches filters.

As a result all action requiring to compute the usage (current usage,
daily_usage...) have to recompute the full usage from 0 every time for
all billable metric having received at least 1 event during the period.
This is highly inefficient especially for charges with filters, has one
filter will lead to one DB query.

With Postgres event store, we are able to use a system with charge
filter granularity, we would like to use the same logic for the
Clickhouse one.

## Description

This PR is a first step to add caching at Clickouse level, it adds a new
`flat_filters` database view. This view will be used on event_processor
side to check filters matching on all events received for organization
using the Clickhouse store.