revert(email): Remove Webhook notification in case of failed (#3813)
## Context

We recently [added this
feature](https://github.com/getlago/lago-api/pull/3654), then [disabled
it](https://github.com/getlago/lago-api/pull/3776). This PR removes it
completely.

## Description

Many customers have organization with failing endpoints (like staging or
dev env) and it's expected because they want to see the webhooks in the
dashboard.

Ideally, the email should be occasional (like Stripe does). This would
require more time and attention that original thought and this is very
low priority.

I believe it's best to revert it completely for now.