feat(alerting): Serialize `customer_external_id` to TriggeredAlerts (#3819)
A TriggeredAlert is saved to DB when the current usage of the
subscription crosses an alerting thresholds.

It's sent over webhooks (serialized once and saved to `webhooks.payload`
column). There is no API or graphql resolver to access triggered alerts.

This adds 2 queries to the serialization. Nowhere in the code we
serialize multiple triggered alert, so no "eager load" to edit.

This is done to help customer easily retrieve the customer to warn.

![CleanShot 2025-06-12 at 09 45
15@2x](https://github.com/user-attachments/assets/d3443021-887b-4f44-8550-5260b78d51e0)