chore(organization): Not null organization_id (#3687)
## Context

This PR is part of the epic to add `organization_id` on all tables of
the application

## Description

The current one is removing the `optional` attribute on all `belongs_to
:organization` relations. In a next step the presence will be enforced
at database level with a not null constraint

The PR also updates the relations from customer and organization table
to make sure it uses the organization_id column.

**NOTE**: Since this change will require all tables to be updated before
the release, a migration guide will be added (See
https://getlago.com/docs/guide/migration/migration-to-v1.29.0 as an
example). The guide will state that customer will have to first run the
[migrations:fill_organization_id rake
task](https://github.com/getlago/lago-api/blob/main/lib/tasks/migrations/organization_id.rake)