chore(db): Squash old migrations into a single file (#3829)
## Context

As we are planing to release the ["not null organization" pull
request](https://github.com/getlago/lago-api/pull/3687), it was decided
to take this as an opportunity to squash the old database migration into
a single migration file, to reduce the number of migrations.

Related to https://github.com/getlago/lago-api/pull/3829

## Description

Lago version 1.20.0 was used as a reference to squash the migrations.

- All migrations from previous versions are merged in a single
`InitialMigration` migration file.
- It uses the SQL format to preserve the orders and make sure the
migration test workflow continue to pass on Github no matter if the
structure comes from previous migrations or from this file.
- It reuse the previous last migration version `20250122112050` to make
sure it will not be executed on instance running v1.20.0 or later.
- It adds a small check to fail running the migration if an upgrade is
made from a version prior to 1.20.0 to the bridge release containing
this change
- All squashed migrations files are removed