fix(no-customer-on-payment):early return when payment method has no c… (#3826)
Whenever we have a `setup_intent.succeeded` with a payment method that
has no customer attached to it.

Stripe does that when setting up payment methods that has some problems
(we don't know exactly which problems are). The `setup_intent.succeeded`
and `payment_method.detached` are fired almost at the same time.

[Once `payment_method.detached` the payment method becomes unusable. 
](https://docs.stripe.com/api/payment_methods/detach)

This raises on
`app/services/payment_providers/stripe/webhooks/setup_intent_succeeded_service.rb`
```
Stripe::InvalidRequestError: The customer does not have a payment method with the ID <pm_ID>. The payment method must be attached to the customer.
```