misc(clickhouse): Rewrite aggregation queries using Arel (#3858)
## Context

After some investigation in our quest to improve performances with
Clickhouse database, it was found that in some condition, when computing
events aggregation, many requests where sent to the Clickhouse cluster
only to retrieve the tables structures.

This is happening specifically when calling `to_sql` on some queries
built using `ActiveRecord`.

These queries are not long to process but they uses some processor time
on the cluster and as such cost money to Lago and slow down the
aggregation processing especially when the cluster is under heavy load.

## Description

As these queries can be easily avoided as they are not required for the
aggregation, since the result relies only on a aggregated values rather
than on a list of record, the decision was made to refactor the
`Events::Stores::ClickhouseStore` to avoid using `to_sql` when writing
queries and to rely as much as possible on pure SQL.

To keep the ability to combine conditions and select, the rework was
made using `Arel` to rewrite the queries.