fix(invoice): Fix race-condition in invoice update jobs (#3851)
## Context

When doing a manual payment, it will update the invoice payment status. When doing so, we're supposed to do a few things asynchronously like sending a webhook.

Unfortunately, the way it is working right now, our server will sometimes process those asynchronous jobs before the database transaction ends, and therefore before we actually update the invoice. We end up with not up-to-date data within those different actions.

For instance, the webhook might be triggered with `payment_status: 'pending'` in the payload instead of `payment_status: 'succeeded'`.

## Description

This commit fixes this issue by ensuring the jobs run after the transaction is committed. This is done by using `AfterCommitEverywhere.after_commit` callbacks.

To reduce the risk of mistakes, two new class methods were added:

- `ApplicationJob.perform_after_commit`
- `Utils::ActivityLog.produce_after_commit`

Doing this, we make the intention clear that the job should run after the transaction is committed.

Test were added for the new methods, and existing tests were updated to properly test that jobs are enqueued after the transaction is committed. To do so, a `expect {}.to enqueue_after_commit` matcher was added.