fix(customer): Fix email validator (#3590)
## Context

Some mailer jobs fail because the recipient email is invalid.
It is due to the email validator bug in case of multiple comma-separated
emails, it was only validating the first one in the list.

This was considered to be valid even if the second email is invalid:
`email1@test-email.com, email2@test- email.com`

## Description

This PR fixes the email validator.