feat(cursor): cursor rules and new matcher! (#3878)
As I'm working on Feature and Entitlement, I added a few new rules to
cursor. I'm extracting them in this PR.

Notice the new matcher, and the rule so cursor use it

About RIPPER MODE, I'm personally not so convinced. The diff is weird
but this is because I changed it from "agent request" to "manual" in the
cursor UI.