fix(aggregator): Add auto-retry logic on SSL error (#3887)
## Context

Many calls to external aggregators are failing with errors like
`OpenSSL::SSL::SSLError`, leading to a failure of caller service. Most
of the time this error is just a transient "bad luck" error and can just
be retries automatically.

The Lago HTTP client is lacking the ability to retry the query
automatically on this kind of error

## Description

This PR adds a new `reties_on` keyword argument on the `LagoHttpClient`
to allow us to define a list of error class that can be rescued and
retried automatically.
For now, it reties 3 times at most before re-raising the error, if it's
not enough, we might extend this approach to allow us to define the
number of retries for each type of error.

This PR also plug this system in the base http_client of aggregator
services