fix(progressive_billing): Avoid checking threshold on terminated subscriptions (#3893)
## Context

We are receiving ` Idempotency::IdempotencyError` for progressive
billing invoices generation on terminated subscriptions.

## Description

We should not generate progressive billing invoices for terminated
subscriptions as the usage will be billed anyway with the termination
invoice.

This PR simply adds a check on the subscription status before proceeding
with the invoice generation.