fix(wallet_transaction): Ensure consistency when creating wallet transaction (#3895)
## Context

An issue has been identifier on the `POST /v1/wallet_transactions` used
to create a transaction on a wallet.

In some situation if the same wallet is being refreshed in a separate
process (job, parallel call to the endpoint), an
`ActiveRecord::StaleObjectError` can be raised when running an update on
the model. The main reason is that the wallet model is using an
optimistic locking approach to avoid inconsistencies.

The error is raised and it leads to an HTTP 500 response error returned
by the API. An other problem appears here: the process to create the
wallet transaction is not transactional, meaning that a wallet
transaction might have been persisted in the database without impacting
the wallet balance...

## Description

This PR:
- Wrap the logic in a single database transaction making sure nothing is
persisted if something goes wrong in the process
- Adds an auto retry logic when an `ActiveRecord::StaleObjectError` is
raised