fix(wallet_transaction): Enqueue BillPaidCreditJob after commit (#3906)
## Context
This PR fixes a regression introduced by
https://github.com/getlago/lago-api/pull/3895

The wallet transaction processing was performed in a single database
transaction. I the process a `BillPaidCreditJob` is enqueued. Since it's
now in the DB transaction, it leads in some situation to an
`ActiveJob::DeserializationError`.

## Description

This PR fixes the situation by enqueuing the job after the the commit of
the DB transaction