fix(payment_request): Prevent payment creation when invoices are payment succeeded (#3910)
## Context

When retrying a failed `PaymentRequests::Payments::CreateJob` (manually
or automatically), we are not checking if the attached invoices are
still ready for payment processing. This lead to a risk of duplicated
payments.

## Description

This pull request adds a new guard to the
`PaymentRequests::Payments::CreateService` to early return when one of
the invoice of the `PaymentRequest` is not flagged as
`ready_for_payment_processing`