fix(specs): Run specs sequentially in CI if the PR is from a community member
Currently, if a community PR is opened, the`KNAPSACK_PRO_TEST_SUITE_TOKEN_RSPEC` secret will not be set. This will cause the specs to fail with an error as we are not able to run Knapsack Pro in this case.

This is not an ideal situation as it will complexify the contribution process.

The specs will now run sequentially in the CI workflow if the `KNAPSACK_PRO_TEST_SUITE_TOKEN_RSPEC` secret is not set.