fix(specs): Run specs using `parallel_tests` in CI if PR is from a community member (#3917)
## Context

Currently, if a PR is opened from a community member, the`KNAPSACK_PRO_TEST_SUITE_TOKEN_RSPEC` secret will not be set. This will cause the specs to fail with an error as we are not able to run Knapsack Pro in this case.

This is not an ideal situation as it will complexify the contribution process.

## Description

The specs will now run using `parallel_tests` (with the `parallel_rspec` CLI) in the CI workflow as a fallback if the `KNAPSACK_PRO_TEST_SUITE_TOKEN_RSPEC` secret is not set.

Note that we don't use `parallel_rspec` to actually run the tests in parallel. By combining `--only-group` with `-n`, we tell `parallel_rspec` to:

1. Split the tests into `KNAPSACK_PRO_CI_NODE_INDEX` (8) groups
2. Run only the `KNAPSACK_PRO_CI_NODE_INDEX` group

So we have the same behavior as Knapsack but without a proper distribution based on timing.