fix(refunds): Fix refunds when payment provider customer is discarded (#3936)
## Context

Performing a refund fails when payment provider customer is discarded.
A refund is performed only on the PSP - the API is called and it
succeeds, but it fails to create the refund record in Lago DB.

## Description

This PR fixes the issue by adding a method that fetches the payment
provider customer even if it's discarded.

`Customers::PaymentProviderFinder#payment_provider_customer`