misc(events-processor): Add new index to speed up in advance charge processing (#3943)
## Context

A slow query was identified on the charge table:
```
SELECT count(*) FROM "charges" WHERE (plan_id = $1 AND billable_metric_id = $2) AND pay_in_advance = true AND "charges"."deleted_at" IS NULL
```

This query is made by the events processor to identify if an event needs
to be billed in advance. The problem was identified because a big burst
of event was post-processed, leading to a pike in CPU usage on the
database.

## Description

This PR only adds a new index to speed up the processing.