feat(entitlement): GQL to create, update and delete features (#3937)
- [x] Introduce new Feature and Privilege type.
- [x] Mutation to update one feature (see note below)
- [x] Resolver to retrieve one feature
- [x] Resolver to retrieve paginated list of features

In GQL, we don't use a hash for privileges but an array a it's more
inline with the GQL way (much better typing).

There is a lot more coming (plan entitlements and subscription
overrides), so this might evolve a little bit before it's fully
released.


### Important note

Notice how the rest API allows you to send partial object to update a
feature while the graphql requires you to send the full object. Given
this payload

```rb
{
  privileges: {
    max: {name: "Max."},
  }
}
```
* API: update or create privilege with code `max` to use name "Max."
* GQL: update or create privilege with code `max` **and delete all other
privileges**.

I believe the API and GQL have 2 differents users (internal vs external)
and they don't have the same requirements.

There are 2 services to update features: `*PartialUpdateService` and
`*FullUpdateService`.

---------

Co-authored-by: Yohan Robert <7434196+groyoh@users.noreply.github.com>