chore(plans): Use update_all to flag plan children for deletion (#3967)
### Expose `pending_deletion` 

When you delete a plan via the API, you get it in the response but
nothing shows that it is deleted or about to be deleted.
Because deletion is async, getting this plan from the API will work as
if nothing happened until the job is processed.

Need openapi and clients update.

### use `update_all` instead of `.each{ it.update! }`

A plan can have maaany children and since we're not calling a service or
relying on any model callbacks, I believe it's best to flag them all at
once.

Introduced here:
https://github.com/getlago/lago-api/commit/69a84700632f760507eee028ede987184d332971