fix(subscription): Ensure `Subscriptions::TerminateService` runs in a transaction (#3965)
## Context

Currently, `Subscriptions::TerminateService` does not execute its logic within a transaction. As a result, if creating the credit note or canceling the next subscription fails, the system may be left in an inconsistent, partially-updated state.

To be fair, I couldn't find a scenario which allows me to reproduce this behavior currently. But the `Subscriptions::TerminateService` will soon also call the `Subscriptions::UpdateService` which will fail if given invalid parameters.

## Description

This commit addresses the issue by wrapping the termination logic in a transaction and using `ApplicationJob.perform_after_commit` and `Utils::ActivityLog.produce_after_commit` to ensure that jobs and activity logs are only executed or produced after a successful commit.

Additional tests have been added to verify this behavior.