test(credit-note): Update job assertions for `CreditNotes::CreateService` tests (#3966)
## Description

https://github.com/getlago/lago-api/pull/3965 ensured that `Subscriptions::TerminateService` wraps its logic in a transaction. Since `Subscriptions::TerminateService` implicitly calls `CreditNotes::CreateService`, the tests for `CreditNotes::CreateService` were updated to confirm that jobs are properly enqueued after the transaction is committed.

Note that I needed to update the `have_enqueued_job_after_commit` matcher so that the `with` chain accepts a block. This is because the job expect to receive the newly created credit note as argument. If we try to pass arguments to `with`, it will be evaluated before the credit note is actually created. So using a block is the only way to verify those arguments.