Feat(subscriptions-list): add search_term to subscriptions and payments (#3898)
## Context

We want to add screens where we can list all subscriptions and all
payments. When we have these screens, we also want to search in them by
search term.
To do this, we added search_term to graphql endpoints and updated
queries
Also to ease the queries, added customer to payments

without direct link from payment to customer, customer is associated
polymorphicly, and the "default" ransack queries should be
`payable_of_Invoice_type_customer_name_cont,
payable_of_Payment_Request_type_customer_name_cont`, which are massive
and ugly. Also if we use this approach, ransack will try to create it's
own join with customers and then will try to use the created join, which
breaks logic if we have to do our own join.
Another way to solve it is to define the ransackers on Payments, but
this way we're creating queries that are related to customer on the
payment + we repeat the code, because ransack search on customer is
already available, but we can't use it.

So I decided to add customer to the payments + we don't have many
payments records yet in the db and it makes sense to keep link between
payment and customer

## Description

- Added customer_id to payments
- Added migrations to populate customer_id
- added constrains on the new column
- Updated all places where we create payments to create payments with
customer_id
- added search_terms to subscriptions_resolver and payments_resolver
- updated queries to handle ransack search
- added exclude_next_subscription filter for subscriptions to handle FE
pulling next subscription themselves