feat(fixed_charges): Add migration for FixedCharge ChargeModel enum (#4000)
## Roadmap Task

👉
https://getlago.canny.io/feature-requests/p/allow-add-ons-to-be-added-to-subscription-invoices
👉
https://getlago.canny.io/feature-requests/p/define-quantities-for-plan-charges

## Context

What is the current situation?

**Option https://github.com/getlago/lago-api/pull/1** - User has to
create a one off invoice alongside the subscription, it will create 2
different invoices.

**Option https://github.com/getlago/lago-api/pull/2** - User can add a
recurring billable metric and use event to have this fee invoice on
subscription renewal, but it won’t appear on the first billing
subscription.

What problem are we trying to solve?

At subscription creation or afterward, there is no clear way to invoice
a fixed fee that is not tied to events, aside from the subscription fee
itself. This fee could be either a one-time charge or a recurring one.

## Description

DB migration to add a charge model enum for the fixed charge model