misc(subscriptions): Refactor `CreditNotes::CreateFromTermination` and its tests (#4025)
## Context

While working on adding refund capabilities to the subscription termination, I refactored some elements of the `CreditNotes::CreateFromTermination`. This refactoring was extracted to make it simpler to review.

## Description

One of the main updates is that I added some documentation of the test cases to make it easier to review and understand. It has the following format:

```
# CREDITABLE AMOUNT CALCULATION
# Unused subscription (16 days)    €16.00
# Previous credit notes            -€10.00
#                                  ------
# Subtotal                         €6.00
# Tax (20%)                        €1.20
#                                  ------
# Total creditable                 €7.20
```