feat(pricing-group-keys): Migrate charges/filters grouped_by to pricing_group_keys (#3853)
## Context

This PR follows the new pricing_group_keys feature that is a renaming of
`grouped_by` and an extension to all the charge models.

The feature was built to allow an easy migration from one key to an
other by keep accepting the `grouped_by` attribute in `charges` and
`charge_filters` when no `pricing_group_key` value exists.

## Description

The current PR adds two migrations to convert all `grouped_by` keys in
`charges#properties` and `charge_filters#properties` field in
`pricing_group_keys`

A next pull request will then clean the code to only rely on the new key