misc: better typing for PrivilegeObject config attribute (#4031)
## Context

We've added the ability to create "feature" object containing an array
of privileges. Those privilege holds a `JSON` type under the config
attribute. This config attribute receive a typing as `JSON` without
details about the keys it can contain.

## Description

This pull request makes the config attribute more explicit to only
include values that it can contain.

---------

Co-authored-by: Julien Bourdeau <julien@julienbourdeau.com>