feat(sso_enforcement): Enforce SSO login across my organization (#3896)
**Feature Link**

https://getlago.canny.io/feature-requests/p/enforce-sso

**Context**

Today, there are 3 authentication methods for users:

- Standard email/password login
- Google OAuth
- Okta SSO, if enabled

However, there are two limitations:

- There is no mechanism to enforce SSO authentication. Even when SSO is
enabled for an organization, users can continue logging in through other
methods, such as email/password or Google. This exposes organizations to
potential security vulnerabilities and compliance risks.
- Only Okta SSO appears in Settings/Authentication, which prevents full
customization of allowed login methods.

**Description**

The goal is to be able to enforce SSO login across my organization (for
security compliance) and to customize which login methods are allowed.

To do that, a new column `authentication_methods` has been added and
populated at the organization level.

We've also added a new field on the `AuthToken` generated by JWT. This
field, called `login_method`, holds the auth constant value defined in
the Organization model. Each auth method saves this information with its
own reference.

---------

Co-authored-by: Romain Sempé <romain@getlago.com>