fix(entitlements): Use Array of privileges instead of hash indexed by code (#4056)
This PR modifies the serializer for features and privileges, as well as
the payload to manage features.
The payload to manage entitlements (attaching features to plans *with
values* is still a hash.

The first thing that made me want to change this was GQL, which was not
so easy to use with hash.
Then, I wrote the open API spec and it was even worse.

### serialized feature / Payload to manage features

```diff
      {
        feature: {
          code: "new_feature",
          name: "New Feature",
          description: "A new feature",
-          privileges: {
-            "priv1" => {value_type: "boolean"},
-            "priv2" => {name: "Maximum", value_type: "boolean"}
-          }
+          privileges: [
+            {code: "priv1", value_type: "boolean"},
+            {code: "priv2", name: "Maximum", value_type: "boolean"}
+          ]
        }
      }
```

### payload to manage entitlements

```rb
{
  "entitlements" => {
    "seats" => {
      "max" => 25
    }
  }
}
```

I think to EDIT entitlements, a hash is a lot nicer. When you receive it
back, it's an array.