feat(subscription): Allow to skip invoice on subscription termination via GraphQL (#4053)
## Context

When a subscription is terminated, whether due to an error or otherwise, Lago automatically generates a closing invoice for pay-in-arrears fees. This inflexible behavior didn't accommodate different business needs where organizations might want to skip the invoice generation.

The API already accepts the termination behavior as a parameter during subscription termination but the GraphQL API did not.

## Description

This update the `terminateSubscription` mutation to handle the `on_termination_invoice` behavior.