integration(netsuite): add plan_code to netsuite payload (#4055)
## Context

A customer wants to add the Lago plan_code of an invoice to a Netsuite
payload. This should be part of the main result payload (not at the line
item level).


## Description

As invoice can belong to multiple subscriptions and therefore to
multiple plans, we'll be sending plans, joined in a string

---------

Co-authored-by: Raffi <raffi@Raffis-MacBook-Pro.local>
Co-authored-by: Anna Velentsevich <annvelents@gmail.com>