feat(wallet-limitations): Add service updates for billable metric wallet limitation (#3973)
## Context

With this feature it will be possible to limit wallet consumption on
specific billable metric fees

## Description

This PR adds needed changes in create and update services for wallets.
Billable metric identifiers are processed so that `codes` are used form
the api side and `ids` are used from the GQL side. Each billable metric
limitation per wallet results in new `WalletTarget` object