feat: change void behaviour and wallet check (#4084)
## Context

When voiding an invoice, the system should recredit only active wallets.
If the associated wallet has been deactivated, the recredit step should
be skipped.

Additionally, when regenerating invoices, we no longer need to recredit
the original credit note, because if we just void the credit note, we
don't want to give anything back to the customer, and if we select to
create credit note, the credit value can be included on the credit note