fix(invoice-preview): Ensure that plan lookup uses only parent plans (#4091)
## Context

When we look for a plan by `code` it possible to find children plan,
which share the same code as parent. For invoice preview that creates an
issue as it could pick up child plan which is not visible anywhere in
the UI and be confusing for users when they got different values from
visible parent plans.

## Description

Ensure that we look for a plan only within parent plans when using
invoice preview feature.