feat(projected-usage): add projected usage data to the current usage endpoint (#3905)
## Context

We offer a endpoint and a UI for the current usage of the customers of
our users. They can see in real time the consumption of each feature for
the current period. However, there is no way to calculate the projected
usage. So we need to add this feature on our platform.
Here is the [spec
document](https://www.notion.so/getlago/Spec-Projected-usage-for-current-periods-218ef63110d280df877bfcb0be4045c4).
Here is the [dive in
document](https://www.notion.so/getlago/BE-Dive-In-Projected-usage-for-current-periods-WIP-21cef63110d280318540fe4310df8048).

## Description

There are many changes in this PR:

- Add a service for calculating projected usage data based on the fees
and the boundary dates.
- Change the serializer for the API to use this new service to calculate
the projected usage values added to the response.
- Change the graphql types to use this new service to calculate the
projected usage values added to the response.
- Made several specs to test the behavior and modified existing ones,
because there were changes to an existing endpoint.